Buildbot: do not run the full test suite if smoketests fail
This is implemented by adding a new build type ("openvpn-smoketest") that can
be enabled on any workers one wishes. The assumption is that smoketest build
run the usual build steps without any special configure options on an
UNIX-like operating system. If the smoketests pass, then the full test suite
is triggered.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>