buildbot-host/buildmaster: Delete openvpn-gerrit-force scheduler
The force schedulers are currently not very useful since
we do not offer any useful selection of changesources. Due
to that the gerrit-force scheduler is specifically useless
since it is a complete duplicate of the normal openvpn
force scheduler. So remove it until we decide to add it
in a useful way.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>